put distance function inside utiities filder
