+++ lib/services/location_service.dart
@@ -2,9 +2,9 @@ import 'dart:async';
 
 import 'package:geolocator/geolocator.dart';
 import 'package:niira/models/game.dart';
-import 'dart:math';
 
 import 'package:niira/models/location.dart';
+import 'package:niira/utilities/calc_distance.dart';
 
 class LocationService {
   final GeolocatorPlatform _geolocator;
@@ -33,41 +33,11 @@ class LocationService {
           List<Game> games, Location userLocation) =>
       games.map<Game>((game) {
         game.distanceFromUser = distance(
-          userLocation.latitude,
-          game.location.latitude,
-          userLocation.longitude,
-          game.location.longitude,
+          lat1: userLocation.latitude,
+          lat2: game.location.latitude,
+          lon1: userLocation.longitude,
+          lon2: game.location.longitude,
         );
         return game;
       }).toList();
-}
-
-// algorithm from:
-// https://www.geeksforgeeks.org/program-distance-two-points-earth/
-double distance(double lat1, double lat2, double lon1, double lon2) {
-  lon1 = degreesToRads(lon1);
-  lon2 = degreesToRads(lon2);
-  lat1 = degreesToRads(lat1);
-  lat2 = degreesToRads(lat2);
-
-  // Haversine formula
-  final dlon = lon2 - lon1;
-  final dlat = lat2 - lat1;
-  final a = pow(sin(dlat / 2), 2) +
-      cos(lat1) * cos(lat2) * pow(sin(dlon / 2), 2) as double;
-
-  final c = 2 * asin(sqrt(a));
-
-  // Radius of earth in kilometers. Use 3956 for miles
-  final r = 6371;
-
-  // returns result in km
-  final resultInKm = (c * r);
-  final resultInMetresRounded =
-      double.parse((resultInKm * 1000).toStringAsFixed(0));
-  return resultInMetresRounded;
-}
-
-double degreesToRads(double deg) {
-  return (deg * pi) / 180.0;
 }
